feat: implement IIncrementalSourceGenenerator instead of ISourceGenerator
