+++ EzId.sln
@@ -24,8 +24,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QKP.EzId.SourceGenerator.In
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QKP.EzId.Benchmarks", "test\QKP.EzId.Benchmarks\QKP.EzId.Benchmarks.csproj", "{3CF55C95-5DF6-4AC4-89E5-810BE6E9B47F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QKP.EzId.CustomTypes", "test\QKP.EzId.CustomTypes\QKP.EzId.CustomTypes.csproj", "{C8396E8B-7A2D-4CA5-834F-D54AED479C0B}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -108,18 +106,6 @@ Global
 		{3CF55C95-5DF6-4AC4-89E5-810BE6E9B47F}.Release|x64.Build.0 = Release|Any CPU
 		{3CF55C95-5DF6-4AC4-89E5-810BE6E9B47F}.Release|x86.ActiveCfg = Release|Any CPU
 		{3CF55C95-5DF6-4AC4-89E5-810BE6E9B47F}.Release|x86.Build.0 = Release|Any CPU
-		{C8396E8B-7A2D-4CA5-834F-D54AED479C0B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C8396E8B-7A2D-4CA5-834F-D54AED479C0B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C8396E8B-7A2D-4CA5-834F-D54AED479C0B}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{C8396E8B-7A2D-4CA5-834F-D54AED479C0B}.Debug|x64.Build.0 = Debug|Any CPU
-		{C8396E8B-7A2D-4CA5-834F-D54AED479C0B}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{C8396E8B-7A2D-4CA5-834F-D54AED479C0B}.Debug|x86.Build.0 = Debug|Any CPU
-		{C8396E8B-7A2D-4CA5-834F-D54AED479C0B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C8396E8B-7A2D-4CA5-834F-D54AED479C0B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C8396E8B-7A2D-4CA5-834F-D54AED479C0B}.Release|x64.ActiveCfg = Release|Any CPU
-		{C8396E8B-7A2D-4CA5-834F-D54AED479C0B}.Release|x64.Build.0 = Release|Any CPU
-		{C8396E8B-7A2D-4CA5-834F-D54AED479C0B}.Release|x86.ActiveCfg = Release|Any CPU
-		{C8396E8B-7A2D-4CA5-834F-D54AED479C0B}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -131,6 +117,5 @@ Global
 		{B9F3D7C5-D9E4-4F0E-B8B5-F3C2E8C6B7A8} = {C882BE14-8067-4DB7-9E00-CBDDCBB980F0}
 		{86389D8E-1575-42E0-BC3F-3D401C8421CC} = {C882BE14-8067-4DB7-9E00-CBDDCBB980F0}
 		{3CF55C95-5DF6-4AC4-89E5-810BE6E9B47F} = {C882BE14-8067-4DB7-9E00-CBDDCBB980F0}
-		{C8396E8B-7A2D-4CA5-834F-D54AED479C0B} = {C882BE14-8067-4DB7-9E00-CBDDCBB980F0}
 	EndGlobalSection
 EndGlobal
