+++ src/QKP.EzId.SourceGenerator/EzIdTypeToGenerate.cs
@@ -1,0 +1,3 @@
+namespace QKP.EzId.SourceGenerator;
+
+internal record EzIdTypeToGenerate(int BitSize, string Namespace, string TypeName, string Separator, string SeparatorPositions, string Length);
