+++ src/QKP.EzId.SourceGenerator/IsExternalInit.cs
@@ -1,0 +1,4 @@
+namespace System.Runtime.CompilerServices
+{
+    internal sealed class IsExternalInit { }
+}
