--- test/QKP.EzId.CustomTypes/QKP.EzId.CustomTypes.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="../../src/QKP.EzId/QKP.EzId.csproj" />
-    <ProjectReference Include="../../src/QKP.EzId.SourceGenerator/QKP.EzId.SourceGenerator.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="true" />
-
-  </ItemGroup>
-
-</Project>
